LP-533 incorrect MCU pin used to toggle inverter for rcvr port on Sparky2
Bug introduced by LP-480. The wrong MCU pin is used for toggling the input inversion. This inverted the serial signal when it should not be inverted.
ibus reception and possibly other serial receiver protocols where not working.

PC4 was controlled, but according to the schematic, PC6 is connected to the xor port.
PC4 is actually connected to the gyro interrupt.

This bug explains another issue I saw from time to time. After a restart of the Sparky2 board, gyro and accelerometer graphs remained locked at 0. Making the board unusable for flight.